Basic UI Set-Up
I created an android project and created three separate activities for a
start-up page, a create-group page, and a confirmation page. Obviously
it doesn't do anything yet, just some buttons and text fields to get us
started. I would image to work with this copy make sure your workspace
in Eclipse is pointed to this repository.