Early Design and Bluetooth Enabling
I began to toy around with making the app more visually appealing,
obviously everything is up for change I'm just trying to understand how
to make this stuff work. Also added permissions for bluetooth access and
made it so when the "Get GROUPED!" button is pressed it will auto-enable
bluetooth if it is not already on.