Changes to UI
First of all I completed the process to ignore the .metadata directory.
With that done I did a little work on the UI. First change is I figured
out how to use pop-ups so those are integrated. Began to make the create
group page a bit more visually appealing. Step to go from here would be
getting info added in UI to objects and then to database.