Added Checkmark Icon
Added a checkmark icon to appear once a field is completed. More
cosmetic stuff but it makes it look a little more professional.... I'm
still just kind of messing around with this stuff. At this point the
only functional button is the name button but refactoring can extend
this functionality to all field buttons.