Got API Key for Google Maps API
It still is having connectivity issues.... as in it won't connect to the
server. But I've got it to a point where fixing it shouldn't be too
problematic. I'm calling it quits on the map stuff for right now so I
can work on merging what we've got so far.